+++ src/helpers/schemas/index.js
@@ -1,3 +1,4 @@
 export { addressDataSchema } from './addressDataSchema';
+export { personalDataSchema } from './personalDataSchema';
 export { signupDataSchema } from './signupDataSchema';
 export { resetPasswordDataSchema } from './resetPasswordDataSchema';