+++ src/index.js
@@ -69,6 +69,7 @@ export {
   zeroPad,
   formatDouble,
   currencyToDouble,
+  formatCpfCnpj,
 } from './helpers/format/string';
 export { default as formatNumber } from './helpers/format/number';
 
@@ -80,7 +81,12 @@ export { default as pxToRem } from './helpers/scales/converter';
 export { validCpf, isEmail } from './helpers/validations';
 
 // Helpers Schemas
-export { addressDataSchema, signupDataSchema, resetPasswordDataSchema } from './helpers/schemas';
+export {
+  addressDataSchema,
+  signupDataSchema,
+  resetPasswordDataSchema,
+  personalDataSchema,
+} from './helpers/schemas';
 
 // Helpers Mask
 export { maskCEP } from './helpers/mask';