+++ webpack.dev.js
@@ -1,6 +1,6 @@
 const merge = require('webpack-merge');
 const common = require('./webpack.common.js');
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+const {BundleAnalyzerPlugin} = require('webpack-bundle-analyzer');
 const WebpackBuildNotifierPlugin = require('webpack-build-notifier');
 
 module.exports = merge(common, {
@@ -11,4 +11,8 @@ module.exports = merge(common, {
     }),
     new WebpackBuildNotifierPlugin(),
   ],
+  watchOptions: {
+    poll: 1000,
+    ignored: ['node_modules'],
+  },
 });